Move da disseminate command into separate module
